chore: upgrade dependencies; compile to CJS
Fix https://github.com/yanivam/react-svg-worldmap/issues/120. The issue seems to
be that we didn't declare `"type": "module"` in our package.json. However, if we
do that, there's the other problem of not being able to import it in CJS with TS
configured with `moduleResolution: NodeNext`. I think a middle ground for now is
to simply compile to CJS.